+++ NEWS.md
@@ -1,5 +1,8 @@
 # systemfonts (development version)
 
+* Added `fonts_as_import()` to create stylesheet urls for embedding of fonts in
+  HTML and SVG
+
 # systemfonts 1.2.2
 
 * Fix compilation on macOS when the obj-c++ compiler fails to pick up the right